pyproject.toml
[tool.poetry]
name = "streamlit-chatbot"
version = "0.1.0"
description = "Streamlit ChatBot with multiple AI models"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.50.0"
langchain = "^1.1.2"
langchain-openai = "^1.1.0"
langchain-anthropic = "^1.2.0"
langchain-google-genai = "^3.2.0"
st-txt-copybutton = "^0.0.3"
st-chat-input-multimodal = "^1.0.4"
authlib = "^1.3.2"
firebase-admin = "^6.5.0"
google-cloud-storage = "^2.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

